SQUASH-ME: Add ERC test scenario for erc-cmd-JOIN
DELETE THIS NOTE: This scenario belongs here because it indirectly
asserts that the changes to erc-cmd-JOIN work as intended.  See note
atop the `ert-deftest' and helper.

The assertion involving the presence of an entry for the current user
in a defunct channel buffer has to do with trying to shift to a
cleaner means of checking whether a channel buffer is subscribed
to (whether it's JOINed or PARTed).  The old means of checking,
basically seeing whether `erc-default-target' is non-nil, depends on
`erc-default-recipients', whose purpose has never been well defined.