[CATCH-UP] Synchronize with patch set
Note: this branch exists for testing purposes and not for code sharing
or merging.  The bug fix continues to be developed as a series of
discrete patches (see email thread for download link).  Commits like
this apply recent changes from the bug set, basically an "--interdiff"
or "--range-diff" from git-format-patch(1).  Non-ERC parts of the tree
are behind master.

This also includes the following from master, but only the hunks that
touch lisp/erc, test/lisp/erc, doc/misc/erc.texi, or etc/ERC-NEWS:

* 2d71fd3b04 Further doc string quoting fixes
* bbf389ea6d Audit quoting the quote character in doc strings